libstagefright: seek notification in case of canceled onAudioCheck
In case of streaming when seek is done to a portion, data
corresponding to which is yet to be downloaded the buffering starts
all over again so the pause is called which results in canceling of
the pending events. In case of an OnCheckAudioStatus getting
canceled, no notification of SEEK_COMPLETE is sent to the upper
layer. Thus, resulting in no response to future seeks.

bug 862180

reviewed on http://git-master/r/48912

(cherry picked from commit 6c26e07890384e04e4485eaaf9f8eb8705d34ddb)

Reviewed-on: http://git-master/r/59026
(cherry picked from commit 67c945855d1b5bf4092628fa946d9adeae81eaae)
Change-Id: Id559361fd69ac159db789982ae9cc233c9f67c35
Reviewed-on: http://git-master/r/64741
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>