SurfaceTexture: order re-use of buffers
SurfaceFlinger with HWC active can set BufferSlots to "free" even when
they have outstanding read fences from display.  When SurfaceTexture has
a buffer count >2, these buffers can be dequeued to applications even if
other buffers were available without outstanding fences.  This can cause
application APIs to block when trying to write to those buffers when
that blocking could have been avoided by dequeueing a different
available buffer.

This buffer dequeues buffers in the order that they are "freed" by
SurfaceFlinger or added.  This is a bit dodgy of a heuristic, but
reduces the odds of dequeing a buffer with outstanding read fences.

This exposed an issue with HardwareRenderer where the
EGL_BUFFER_PRESERVED bit was lost when the surface was destroyed and
re-created.  This caused glitches on the unlock screen and was also
fixed.

Bug 908098

Change-Id: I6dcf48de5b092753a2db910f23f5334f3b4214b9
Reviewed-on: http://git-master/r/66005
Reviewed-by: Jonathan Mccaffrey <jmccaffrey@nvidia.com>
Tested-by: Jonathan Mccaffrey <jmccaffrey@nvidia.com>
Reviewed-by: Dan Willemsen <dwillemsen@nvidia.com>